Create asc.js (fixes: #8) (closes: #11)
With the necessary dependencies (or within `nix-shell -A js`) running

    make asc.js

will create `asc.js`. This java script file can either be used with the
browser or with node.

In the browser, it exports a top-level object `asc` with a field
`compile`.

In node, you can `require` it to load it, and then `asc` is a member of
the resulting module.

`compile` is simply string-to-string for now, and errors are written to
the `console.log`.

Example session:

	$ nix-build -A js
	/nix/store/h6fpriplws9z4i2z3bfi3vs2kyad0jr5-js
	$ node
	> var m = require('/nix/store/h6fpriplws9z4i2z3bfi3vs2kyad0jr5-js/asc.js')
	undefined
	> wat = m.asc.compile('assert(false)');
	'(module\n (type …