Run dsh in the test suite
this requires including it in `default.nix`. And with “it”, I mean
dev-in-nix, because that is where the derivation for `dsh` is defined.
But because `dev-in-nix` is a private repository, I cannot easily
include it via `fetchGit` (in which case nix would pull it behind the
scenes). So I am including it as a git submodule :-(

But hey! We ran “hello world” on the hypervisor!