Create an “extended module” type with additional custom sections
to no longer thread the `custom_sections` value around separately.

This also drops the `compileWat` function. The text format does not
allow for custom sections, so it is not faithful, and should be used.
If textual representation is wanted, then the `.wasm` can be converted
to `.wat`. Especially after implementing the `name` custom sections,
this will actually yield better results.