Check use-before-define in the type checker
Turns out this is super simple once we have the annotated information
from `freeVars.ml`, and does not even have to distinguish the various
forms for declaration.

This fixes #74.

All tests go through, so it does not seem to be overly restrictive. It
could certainly need more tests and some eyeballs if it is actually
sound.