Make duplicate field error fatal
the function gather_field_exp is run thrice; not a good place for a
non-fatal error.

This could be refined to make sure the check happens only once, either
by passing down a flag to exactly one invocation of `gather_field_exp`,
or moving the check to a separate, explicitly used function alltogether.