Expose less internals in typing.mli
this has been nagging me for a while: typing.mli was more complex than
it needed to be. So this simplifies this a bit.

If this change is accepted in spirit, I would probably turn `scope` into
a record. And maybe make it field of the internal `env`, rather than
duplicating the field.

Also, given that the result of `check_prog` is always just `adjoin`’ed
to its argument, why not do that in `typing.ml` and even remove
`adjoin_scope` from the interface.