Desguar all objects to NewObjE, and compile it properly
if we have to do it for some, lets do it for all! So in `desguar`, we
now replace `ObjE` by `BlockE { LetD … ; Var D ; LetD o NewObjE }`,
properly implementing the idea of an object as a bunch of closures.

Object class functions are also desugared here. The class identity is
lost. In light of getting rid of `instance-of` (#81) that seems fine for
now.

Actors cannot be compiled that way, so they survive untouched (as
`ActorE` and `ActorClassD` to be explicit).

The compiler now implements `NewObjE` so that `{x = y}` makes `x` an
alias for `y`, so that mutable fields stay mutable.

This fixes the semantics of `test/run-dfinity/async-obj-mut.as`. Yay!

For now, *all* actor fields have this level of indirection.  A further
refinement may restrict that treatment only to mutable fields.

Small wart: If the compiler finds that a mutable variable is not
captured by a function, it will allocate it on the stack. If such a
variable is put in `NewObjE`, the content will be put in a new box. This
is fine for the output of the desguaring: If it was not captured, the
local `x` is not referenced after the object has been created.
But this is shaky ground, and may need to be revisited (probably by
beefing up the `AllocHow` algorithm that determines whether to
heap-allocate a variable).