Copy Wasm.Ast and Wasm.Types into this repository
because we need to include multi-value support.

I tried to get rid of the wasm reference interpreter dependency
alltogether, given that we really only needs the AST (we already have a
copy of the serialization module), and it would be nice to have a single
`Source` module, for example, but we are using a lot of stuff from
`Wasm.Lib`, e.g. for value types.

Once multi-value support is officially supported by the `wasm` library I
am happy to undo these changes and start using the official version
again.