Record type of overloaded operators
the type infernce code knows the precise type that it has determined is
compatible with a given operation (`t` in `typing`). But previously,
this was not stored anywhere, and interpreter and compiler had to make
educated guesses based on the type annotation of the arguments and/or
the return values. This is brittle, due to subtyping, and led to bugs.

Therefore, we not simply store the type within the constructor. This
fixes #122.