Traps in `dvm` are now silent, as they should
thanks to https://github.com/dfinity-lab/dev/pull/408

But in order to keep the test suite useful, `run.sh` now automatically
appends

    print("Top-level code done.\n")

to all test cases (in `--dfinity` mode). This way, a trap in the
top-level code will be observable in the `.dvm-run.ok` file

It does not help with, say

    ignore(async {assert False});

so tests that want to check the absence of traps need to somehow make
that visible.