Tidying interpret.mli
it turns out that outside `Interpret`, all fields of `Interpret.env` are
always empty, with the exception of `val_env`. So it is silly and
confusing to even expose that in `interpret.mli`. No user of `Interpret`
needs to know about the `labs`, `rets` or `async` fields.

This makes the interface a bit more narrow.