An interpreter for the IR
this is currently a pretty straight-forward copy of `interpret.ml` to
the new IR. The test suite ensures that we get identical output with
either `Interpret` or `Desugar`+`Interpret_ir` (and I checked that the
test suite would report the differences if there were any).

This is still uses the CPS style, and I think this is unavoidable if we
want to be able to interpret our code at any stage of the pipeline,
including before and after the await-translation. Also, for code resuse
in `Value` (which assumes a CPS’ed function type).

Fixes #141.