An interpreter for the IR (#143)
* Tidying interpret.mli

it turns out that outside `Interpret`, all fields of `Interpret.env` are
always empty, with the exception of `val_env`. So it is silly and
confusing to even expose that in `interpret.mli`. No user of `Interpret`
needs to know about the `labs`, `rets` or `async` fields.

This makes the interface a bit more narrow.

* An interpreter for the IR

this is currently a pretty straight-forward copy of `interpret.ml` to
the new IR. The test suite ensures that we get identical output with
either `Interpret` or `Desugar`+`Interpret_ir` (and I checked that the
test suite would report the differences if there were any).

This is still uses the CPS style, and I think this is unavoidable if we
want to be able to interpret our code at any stage of the pipeline,
including before and after the await-translation. Also, for code resuse
in `Value` (which assumes a CPS’ed function type).

Fixes #141.

* Keep inner region when desugaring AnnotE

this makes the output from the two interpreters match in more cases
(otherwise the error messages would have different source locations.)

* Test suite: Avoid leaking time-stamps into the `ok/*.diff-*.ok` files.