Remove the concept of a shared objects
previously, objects might have had a class identity that is not valid
across different actors, so we distinguished between shared and
non-shared objects. With `is` removed (#145), we no longer that
distinction; an object is a subtype of `Shared` if and only if all
fields are (same as for tuples, as one would expect).

This removes a bunch of code. It might be that Andreas could have
implemented this faster on his own than reviewing this PR, but I wanted
to sieze the opportunity to learn more about the type chekcker etc.

It is one of the clean-ups menionted in #81 (there are more left though).

It fixes #103.