Base IR.typ_bound on `con`, not `var`.
Previously, `IR.typ_bound` would mirror `Syntax.typ_bound`, with a `var`
instead of a `con`. But post-typechecker, we only want to work with the
stamped `con`. Previously, this was hidden away in the annotations
(misusing `Type.typ` as `Con.t option).

This patch cleans this up a bit,
but introducting a `Type.open_bind` that binds a `Con.t` to its bound,
and using that as part of `FuncD`.