Bundle the kind with the con
The benefit is that many passes no longer need to keep a `con_env`
around, and many operations in `Type` are now always available, in
partiuclar `Type.as_*_sub`, `promote` etc.

The downside is that, because of the multi-pass approach in `typing.ml`, the
annotation needs to be a `kind ref`.

I tried using a `promise`, but it seems that certain things are run more
than once in `typing` (the same has caused issues with recoverable error
messages before). I am not sure if this is avoidable.

If maybe the initialization can be made a bit cleaner in `typing`, then
I think this will improve the life of those working on the `Ir`
noticably.