Refactor: No typ_binds in `Type.Abs`
previously, there were three uses for the kind `Type.Abs`:
 * type definitions using `ClassD`, because a class instance is a
   subtype of the like type of the the class
 * bounded type parameters
 * type definitions using `TypD`, in the `pre`-phase of the type
   checker.

With nominal typing gone, `ClassD` should introduce a type equality, no longer
a type inequality, so the first use case is gone.

In the second use case, the parameter list was always empty (type
parameters do not have themselves paramters, i.e. we do not support
higher kinds).

Assuming that it is ok in the last use case to use `T<A,B> = Pre`
(intead of `T<A,B> <: Pre`) during the `pre`-phase of the type checker,
we can simplify the types by removing the parameter list from
`Type.Abs`.

This also removes the case
```
    | _ when con1 = con2 ->
      rel_list eq_typ env rel eq ts1 ts2
```
in the `rel_typ`, which might have been a bug before we got rid of
nominal typing.