Bundle the kind with the con (#163)
* Bundle the kind with the con

The benefit is that many passes no longer need to keep a `con_env`
around, and many operations in `Type` are now always available, in
partiuclar `Type.as_*_sub`, `promote` etc.

The downside is that, because of the multi-pass approach in `typing.ml`, the
annotation needs to be a `kind ref`.

I tried using a `promise`, but it seems that certain things are run more
than once in `typing` (the same has caused issues with recoverable error
messages before). I am not sure if this is avoidable.

If maybe the initialization can be made a bit cleaner in `typing`, then
I think this will improve the life of those working on the `Ir`
noticably.

* Simplify Ir.TypD

the `kind` is already in the `con`, so no need having it here.

* Make the representation of con.kind private

so that the choice whether it is a `promise`, or a `ref`, or something
else, private to `type.ml`.

I tried to make it more promise-like, by trapping when the kind is set a
second time to something different, but it seems that `k = k'` can loop.

* Make set_kind check that the kind is not set a second time

with a different value. A bit like a promise, but more compatible with
`typing.ml` running `infer_dec_typdecs` multiple times.

Not sure if it is worth the bother, though.

* Address nit

* Refactor avoid to take a `con_set`, not a `con_env`

* Drop con_env argument to Type.avoid (use kind inside)

* Remove `con_env` from `scope`

In `typing.ml`, we replace it with a `con_set`, so that we can still call
`avoid`.

In `check_ir.ml` we simply drop the field.

* Actually, get rid of `Typing.con_set`

it is always the range of `typ_env` anyways.

This simplifies a few more functions.

* Delete definition of `Type.con_env`

* Make `kind` a field and type parameter of `Con.t`

* Write an .mli file for `Check_ir`

which uncovers some dead code.

* Address nits

* Better name for the abstract annotations on cons

* restore con envs

* replace con_env by con_set in typing.ml

* - simplify IR con_env to have type ConSet.t
- we retain the con_env field name as rossberg suggested
- make some add and unions, disjoint_add and disjoint_union to stress
  the invariant on con_env and catch more errors

* inlined to_avoid (fixing nit)

* "fix nits and clone constructors in async tranformation to avoid side-effecting the source program (and whatever else is referencing cons)"

* rename to_avoid 2 cons

* only call set_kind once with a proper definition; for now, assert kinds are equivalent

* cleanup kind setting