Change how async deals with stamps and kinds
to allow for the (I think) desirable invariant: Same stamp ⟺ same ref.

Previously, the pass would clone all type constructors, even those not
defined in the current `prog`. It worked, but I find it fishy: Why cone
`Int` when `Int` is not defined here?

So this patch proposes to do something different, namely leaving
already defined cons alone (rationale: if they are not defined here, we
don't get to change them).

Eventually, if we want to do an `async` translation on two separate
program fragments, then we will have to pass the `con_renaming` down;
everything else is a hack that does not generalize to more complicated
type-changing passes (e.g. those where the type change depends on the
actual code, such as inlining or unboxing).

Oh, and it would be cleaner to separate `t_con` into `t_con_def`
for binding sites (where we always clone) and `t_con_def` for occurrence
sites (where we only lookup and fail if it is no there), and the
invariant that the domain of `con_renaming` is always the in-scope cons.
I did not do that because it is somewhat tedious with recursive
declarations, and the current code works. But we might have to do that
eventually. (It is a pattern that is very familiar to me from GHC.)