AST58: New Actor IR
An actor in the IR is now a bunch of “normal” declarations, follows by a
mapping between exported label and local id.

I dropped the `self` field from `IR.ActorE`. The compiler has a hack to
support that particular idiom; I will tackle general actor value capture
next.

The current implementation of `ActorE` in the backend is most general:
It assumes that the declarations decare some `funcref` (as they
necessarily do, given their types), stores this in a static locations,
and then exports a small wrapper function that calls this `funcref`.

In general, that is the right thing to do (the funcref could be captured
from the enviornment). But in most cases, we just want to call the
function directly.

There is also one test case that makes v8 throw a nasty exception, and
it cannot be minimzed separately. Not sure what I am doing wrong. Will
report that against `dev`.