Refactor Ir.BlockE
it became clear that we always want the declaration list in a `BlockE`
to end with an `ExpD`. So lets make that explicit.

This also allows us to drop the type annotations on `dec`.

Because of the needs of the REPL, an `IR.prog` is now also a
`(dec list * exp)` pair.