Introduce IR.FuncE (AST-26) (#196)
this turned out to be a bigger change, uncovering a few dirty spots in
`Check_ir` and other places.

This codifies also that, in the IR:
 * All `BlockE` expressions end with an `ExprE`
 * Only `ExpD` declarations return their type (this is differs from Source)

In `Check_ir` checking of `FuncE`, we make no assumption about the type annotation
(which may be a supertype). Instead, reconstruct the function type from the arguments of
`FuncE` and compare it to the annotation. HT to Claudio for spotting this.