Extend testsuite to allow for FileCheck assertions
[FileCheck](https://llvm.org/docs/CommandGuide/FileCheck.html) is a tool
created by the LLVM folks to embed textual assertions about the output
code in test cases. This comment sets up our test suite so that we can
put `CHECK` directives into the test files, and `run.sh` will use
`FileCheck` to see if the resulting `.wat` file matchs the specs.

I am using this in one example to assert that the mutually recursive
functions in `mutrec2.as` are compiled to direct calls (and not
closures).

This requires you to have `FileCheck` in your PATH. The easiest way of
doing that is to run

    nix-env -i -f . -A filecheck