Create a pass that inserts calls to serialize/deserialize
which requires a new primitive type for `elembufs` (the current
representation of data). This prepares for type-driven serialization, by
isolating the points where serialization happens.

Serialization is still based on the dynamic type.

The failing `asyncreturn` is because of the polymorphic message type
there, which we do not support.

There is some extra reference boxing/unboxing introduced now, which I
will get rid of separately.