Ir: FuncE takes a list of names, not a full pattern
if we have n-ary functions, then really a function definition ought to
have n named arguments, rather than a full pattern, and the desugarer
should move the pattern into a `LetD`, if necessary.

I did this in order to unbox function arguments more easily, but I think
it is generally a useful direction.