Rule out messages with abstract types in the type checker
My work-in-progress implementation of type-based serialization
`assert`s `false` when it encounters such, but it is much nicer to just
rule them out statically.

I updated some tests that were using abstract types in messages, but
were (I believe) testing something else to use concrete types.

The test case should outline what it allows and what it forbids.
Interesting design decisions:

 * I only check that parameters, arguments and return types are concrete
   upon a shared function definition and shared function calls, but not
   within any type annotation.
   The main purpose is to still allow type definitions to have abstract
   types in these places, as long as they are made concrete when they
   are used. This allows

       type X<B> = shared B -> ()     // abstract type in arguments
       func foo ( f: X<Int> ) = {f 5} // now a concrete type

 * I do not explicitly forbid shared functions to have type parameter
   lists. Because of the other check this is only possible if the type
   parameters are not used (i.e. are phantom), so doesn’t matter in
   practice. But there is also no hard reason to forbid them right now.

We *could* make `is_concrete` return `true` for functions and actors, if
we want, because their representation is independent of the types of the
arguments/fields. This is a neat observation, but probably more
confusing than relevant, so I am *not* doing that.

The implementation of `is_concrete` uses the same memoization technique
like `rel_typ` to deal with recursive types, and may diverge when there
is non-uniform recursion (but so will the code generation for
serialization).