Small IR refactor: Make FuncE return type explicitly n-ary
I think that every use of `Type.as_seq` is a bit fishy: Once we are in
the IR, we really should think in terms of n-ary types as far as
possible. So let’s store the list of return types in `FuncE`, making it
match the `Type.Func` constructor more closely. `Check_ir` is
simplified, which is a good thing.

This does not completely eliminate `Type.as_seq` from the IR phases
completely; especially the `async` translation (and hence parts of
`construct`) use it to flatten the return values of `T.Async`. Maybe
`T.Async` itself should take a list of arguments, so that we can deal
with that in the desgurarer, where we already deal with fuction
arguments and return values.