Print message before trapping, on DFINITY mode
previously, in a large number of possible error conditions, whether they
are benign (array index out of bounds) or internal (type error,
unimplemented code) we just called `Unreachable`, causing a trap. But
this is really not helpful to our early adoptors.

So in `DFINITY` mode, before calling `Unreachable`, print a helpful
error message via `test.print`.

I made the strings all static, so that multiple calls with the same
error message do not have to allocate static spaces multiple times. (It
currently still does, but making sure every static sting is only added
to memory once is easy to do, and independent of this change.)

Some cleanup of `compile_lit` along the way. Also I prefer `let open Foo
in` over `Foo.(…)` when `…` is a multi-line expression with control flow
and it is the last expression of the current function anyways.