Fix the product-like IR subtyping rules for good.
For `ObjP` now we revert back to the previous code, with the ordering
fix added.

For `TupP` we now require the scrutinee type to be a subtype of the
pattern's type.

If we wonder why the tests ever passed with the latter IR-checker bug,
the reason is that the component's types were already pushed towards
the most refined subtype, thus the check found both being equal, i.e.
subtypes both ways.