Implement import statements
with the semantics of
https://dfinity.atlassian.net/wiki/spaces/AST/pages/147357892/ActorScript+library+import

This is the simple variant where we translate them between parsing and
typechecking into a big block, with (internal) variable names for the
files.

Things that probably should be done (but maybe in a separate PR):

 * [ ] Implement the check if the contents of a file is static
 * [ ] If we don’t want recursive modules (@matthewhammer?),
   then we can type-check them in sequence, likely improving error
   messages and code
 * [ ] If we do want recursive modules (@matthewhammer?), then
   we should probably still refactor this so that we don’t piggy-back
   upon `LetD`, and instead have a data type that describes “list of
   improted files plus body). Probably also worth having a separate
   value environment indexed by filename, to not conflate them
   too much. Would remove the need for `id_of_full_path` and
   `is_import_id` in the front-end.

The file `pipeline.ml` is still a mess, I think I did not make it a
bigger mess. I tries some clean up, in particular:
 * Lexing, parsing, resolving always applies to a single file (or
   string)
 * loading, checking, interpreting and running always applies to a list
 of files