AST-12 add debug_show (#302)
Besides exploring type-driven code synthesis, this provides a useful
debugging tool that @matthewhammer was desperate for.

Multiple pieces work together:
 * The type checker ensures that every call to `debug_show` has a
   showabletype as the type instance for `A`.
   Showable types are defined by `Show.can_show`.
 * The interpreter interprets `ShowE` using `Show.show_val`
 * There is an IR-to-IR pass that replaces `ShowE x` with a call
   to a monomorphic function that implements it for `t` properly.
   It is compositional and multiple calls to `show` in the same module
   share the code for common subtypes.

This represents `debug_show` as its own AST node. I still think that
it is conceptually an `UnOp`, and it would spare us lots of extra lines
in AST passes, but a majority of people thought that it is not really an
unary operator.

Note that `debug_show` for `Text` does not escape the string.