Implement import statements (#307)
with the semantics of
https://dfinity.atlassian.net/wiki/spaces/AST/pages/147357892/ActorScript+library+import

This is the simple variant where we translate them between parsing and
typechecking into a big block, with (internal) variable names for the
files.

The file `pipeline.ml` is still a mess, I think I did not make it a
bigger mess. I tries some clean up, in particular:
 * Lexing, parsing, resolving always applies to a single file (or
   string)
 * loading, checking, interpreting and running always applies to a list
 of files

This comes with a syntactic check for “static expressions”, to be used here and
probably in #175. The check is a bit uninformed yet about what it should allow and
what not, and may need to be tweaked.