disambiguate shorthand variant parsing
Okay, this is understood. It is related to the juxtaposition call syntax: When `#foo` can be an `exp_nullary`, then `#foo #bar` can be a `VariantE` or a `CallE`, and menhir resolves it the former way. That is why we didn't get an error on the testcase (`variants.as:52`). So we were lucky, but I have following plan: Exclude the shorthand form (lone `#foo`) from becoming a postfix expression (`exp_post`). We lose
```
```
which are all nonsensical on semantics' grounds (and would be rejected by the typechecker anyway).

Then we have to re-add the shorthand form for `exp_un` (i.e. the only consumer of `exp_post`) in a verbatim way. If you have a better idea how to salvage the situation, please tell.