Type-check imported libraries separately
previously, we syntactially combined them into one big AST, and
typechecked that, but that was obviously just a make-shift solution.

So now `pipeline.ml` is driving the process of loading and type-checking
each imported file separately.

This led me down a rabbit hole of rewriting most of `pipeline.ml`. Some
noteworthy changes:

 * The public interface of `pipeline.ml` is much simpler now. (Yay)
 * `Syntax.prog` stores the filename in the note field; this way we have
   pass less `name` parameters around
 * Because we now also type-check the files on the command line
   separately, there is no longer recursion between these.
   I deleted the `samples/app` example, which was highlighting exactly
   this particular recursion. Sorry, @crusso.

Still not all is rosy.