Move the Syntax.id_of_full_path into the desugarer
and thus out of the frontend (Typing and Interpret).

Also make sure that we do not load a library that we loaded already
before in the repl.

This also adds the beginning of a test suite for repl tests, because
some there are some features there that we can’t test otherwise.