AST-80: Public actor fields must be functions
otherwise every message would go through a second, forwarding message
send, which is bad. Or we need to do some best-effort optimizations, but
that is also not nice. So let’s require a function here (the user can
still eta-expand they really have to, so no loss of expressivity).

The actual changes to the IR and backend will come in a separate PR.