Encoding and decoding of extended modules
this PR is pulled out of my work on linking #349, to make reviewing
easier.

We already had a copy of the module encoder from the Wasm reference
implementation to add support for source maps. This pull request

 * moves the code that encode the custom sections we need (names, plus
   the DFINITY specific ones) from a separate module into this modules,
   which is nicer (less duplication of the decoding infrastructure).
 * adds the dynamic linking section to our notion of `extended_module`
   (but not encoding yet, might be added separately)
 * adds an decoder for these extended modules, again copying
   the reference implementation code and adding the necessary features.

Not all parts are exercised here yet (in particular not the decoder),
but I figure this is still better than a huge code dump with the whole
linking story.