The Wasm linker
this is part of #349, and adds the linker module. It does not use it in
`asc` yet, but exposes it as a separate binary (`as-ld`) for testing,
and starts a small test suite.

This builds on top of #376 and should probably be reviewed after that.

The code in linkModule.ml is not yet the epitome of nice code structure,
and I will improve the code as I keep working on it. But I want to get
the functionality out and tested.