The C RTS part
again a part of #349: This PR pulls in the C files that will be part of
the RTS, and add support for building them with nix.

There is a README with build instructions, and some technical details in
comments.

So far a single `.c` files for simplicity, but expanding this to
multiple files should not be a problem.