Backend: link against as-rts.wasm
this is the next (and probabaly second-to-last) chunk of work from #349:
Linking as-rts.wasm in the backend.

The `--no-link` flag is mostly for debugging; the unlinked module
isn’t yet useful (but might be if we start to do dynamic linking on the
chain).

Still left on #349: GC handling for big ints.