Backend refactor: Move ActorScript name environment further down
`compile.ml` is mostly split into two parts:
 * the back-end leaning parts: Heap objects, functions on them etc.
 * the frontend-leaning parts: Traversing the AST, keeping track of the
    variable bindings etc.

But so far we had one environment for both aspects (e.g. storing names
for WebAssembly locals; keeping track of where AS variables are stored).
This has bothered me for a long time, and I finally pulled the band-aid:

* `E.t` no longer has `local_vars_env`. In fact, `E.t` is only concerned
  with the bookkeeping needed to prodcue Wasm.

  Also `env : E.t` now only changes when going into a new _Wasm function_,
  no longer with every ActorScript scope. For some local `go` function,
  it no longer needs to be passed around.

* There is a new environment `ASEnv.t` which only contains the
  ActorScript name and label environment. It is defined much futher
  down, and passed around separately by `compile_exp` etc.

  Only the `ae : ASEnv.t` changes when entering a scope, so only this
  needs to be threaded through patterns or declartions etc.

This also unbreaks the unholy nested type recursion through `varloc`. I
expect that I can simply the `deferred_loc` story now that this is
distangeled.

In theory we should be able to do `open IR` before the second part, but
we are not quite there yet.

Bitwise identical Wasm output with this change, according to
`./compare-wat`.