Backend: Consider more blocks as top-level (#445)
* Consider more blocks as top-level

this was previously just a typo in `compile_top_lvl_expr`, I think.

If the top-level contains further blocks, then they can also be compiled
as top-level (yay, less dynamic allocation, more calls to known
functions).

This uncovered a bug around `AllocHow.decs_top_lvl`, which needs to know
which variables are captured in the body to make the right decision. So
feed that through. Not particulary pretty, but at least correct.

* A similar bug existed in AllocHow.decs_local