Backend: Allow `actor class foo() { .. }` as last expression
this is in anticipation of #400.

It extends the prevoius hack in the backend, and does not include the
necessary changes to the front end for proper error handling. But it
finally allows the produce exchange to use `import` for everything, no
longer relying on the concatenation of source files (relevant for #447).