An initial draft for the IDL-AS mapping spec
I tried to first list goals (including reasonably formal ones), maybe
lets focus on them first.

I also want this to be a blueprint for other languages. In particular,
in how we handle types that exist in one language but not the other
(`Char` in one direction; `float32` in the other).

Not done yet: The tooling interface and workflow specification (should
it be in this document or somewhere else?)