IR: Use Type.lab instead of Name (#451)
In contrast to the source, the IR has a clear separation of what are
variables (can be renamed) and what are record labels (must not be
renamed). E.g. thanks to @crusso’s `NewObjE`.

And because the label in, say, `DotE`, directly refers to the
label in a type, it makes sense to use `Type.lab` instead of `Ir.name`
here.

This allows the backend to move the `open Ir` much further down, making
the structure of that file more obvious.

Currently, `Type.lab` is an alias for `string`, and it might be
desirable to make that a new type (as `Ir.name` was before), but that
can happen separately.