Draft source reorganization
this is a proposal, based on the idea of taking some “roots” (types,
source, ir, wasm-stuff), and then grouping the modules by which subset
of these roots they need (or should) depend on.

This will reveal some layering violations and require some cleanup. For
example, after this reorganization, the Interpreter will no longer be
able to just peek at `Flags`, and they need to be passed in. This is a
good thing.

Also, the source map generation currently wants the source of the
prelude. This is also not nice (and it should just somehow get all the
sources and include them in the source map anyways).