Move call_conv into Type
there are many modules that were using `Value` only to get funtions
related to the calling convention, which really is more closely related
to the type of functions (it is a project of the function type, after
all). Moving stuff into `Type` distangled the module dependency graph a
bit.

Might help with #843, in particular separateing `Value` from `Type`, as
suggested by Claudio in
https://github.com/dfinity-lab/actorscript/pull/483#discussion_r291564774