Source reorganization (first big step) (#483)
It turned out to be pretty tedious, and at some point I took some
shortcuts, so there are now edges in the dependency graph that we do not
actually one. For example, some definitions from `Syntax` permeate the
`Ir` world as well.

Instead of fixing everything in one go, I propose we merge this, and
then take further cleanup steps.

Some clean-up steps that I did take:
 * split `effect.ml` in the Source and the Ir part.
 * Move data types for operations in to `operation.ml`, as they
   are shared between Syntax and Ir. There needs to be done more in that
   direction.
 * Move `Arrange.prim` to `Arrange_type.prim`.
 * Splitting the `prelude.ml` into the text of the prelude, and the
   interpreter semantics of the primitive functions.

The decision whether we `open Libname` or not was not taking with too
much considerations yet, and can be refined further. Also which
libraries should export their set of modules unwrapped (bad practice
according to the dune documentation).

The `flags` library is also a stop-gap measure, as is the `prelude`
library.