Do not read Flags.source_map in customModuleEncode.ml (#510)
one could tell that this code was written by a Haskell programmer: The
Source map is generated in any case; the flag was only controlling
whether to stringify it or to return `""`. So let’s just always generate
the source map for now; the main function decided whether to actually
write this out to file or not.

At some point we can make sure that `customModuleEncode` doesn’t even
create the source map if we don’t want it, but that’s an optimization
for later.

Together with #508 this means we can remove the `flags` library!