Use an explicit NumConvPrim instead of string-typed prim names
this extend the IR AST with that node. IR checker not (yet) extended.

The transitition from Source (with untyped prims of arbitrary string) to
IR (with a more structured AST) is a bit, eh, bumpy. But that might be
fine, as this bumpyness only applies to code written in the sacred
prelude.