asc: -no-multi-value (#607)
For some use-cases (e.g. processing the compiler output with analysis
tools) it is useful to avoid the multi-value extension.

This flag provides mostly transparent wrappers that put multiple values
in statically allocated globals (at most 10...) and pull them off again.

If the flag is not used, the code is unchanged.

Also test the -no-multi-value flag in CI, as it is bad style to have
code paths that are not tested.