IR.FuncE: Check calling convention against the parameter and return type list (#648)
ever since
4efd91586f3cb3d5b8354186494a062353541353
and
4e67f89dac0a4dc3904374d206e29f3848abf222
the parameter list and return type list of `FuncE` should match the
Calling Convention field. This checks this fact.

Maybe it is time to get rid of the calling convention, now that we have
almost all relevant data in the other fields, moving `FuncE` closer to
`FuncT` again.

I wondered if this check breaks due to the way we put `[AyncT t]` in the
return type list for shared functions, but the test case I added says we
don’t. Maybe we just don't do n-ary returns in async these days.