Do not return unexpected unit tuple
This is left-over code from when `ret_tys` was still meaningful in the
backend. With the async pass back it, `ret_tys = []` is always true, so
we cannot detect one-shot functions this way.

It seems that the system does not actually check yet if `reply` is
actually called, so if we just don’t call it, that’s fine too. That is
of course a bug in the system, but it helps us for now.

Looking forward we have to do one of these:
 * Get rid of one-shot functions (#661) (my favorite)
 * Reliable detect them in the backend, and reply there.
 * Detect them in the async translation, and create a fake reply there.
 * Add proper support for them to the system

BTW, the `idl-tuple` output shows that our treatment of functions
returning multiple values is curretly broken, because the pair hides
behind the `async` constructor it is not treated as a n-ary return. I
hope that we soon get to refactor our function types to get rid of this
“administrative `async`” in the return type, which would hopefully fix
this.